socialai/WebApp-API-develop/requirements.txt
SQLAlchemy==1.3.24
Flask==1.1.1
Flask-SqlAlchemy==2.4.1
Python-dotenv==0.10.3
Flask-Migrate==2.5.2
Python-Jose==3.0.1
requests==2.26.0
Graphene-SqlAlchemy==2.2.2
Flask-Graphql==2.0.0
elasticsearch==7.0.5
flask-cors==3.0.8
elasticsearch-dsl==7.1.0
boto3==1.20.11
graphene-sqlalchemy-filter==1.10.0
graphql-core==2.3
pytz==2019.3
funnelmodels==0.9.34
twilio==6.60.0
python-dateutil==2.8.1
phonenumbers==8.12.35
cryptography==35.0.0
flask-cors==3.0.8
Flask-SocketIO==5.1.1
python-engineio==4.3.0
python-socketio==5.4.1
pyfcm==1.5.4
